% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTypeMeasures.R
\name{availableTypeMeasures}
\alias{availableTypeMeasures}
\title{Display the names of the measures used in CV for different families}
\usage{
availableTypeMeasures(
  family = c("all", "gaussian", "binomial", "poisson", "multinomial", "cox",
    "mgaussian", "GLM")
)
}
\arguments{
\item{family}{If a family is supplied, a list of the names of
measures available for that family are produced. Default is "all", in which
case the names of measures for all families are produced.}
}
\value{
If `family = "all"`, a list of names of measures that can be used
in CV for each family; otherwise, a vector of names of measures that can
be used for the family passed as the parameter.
}
\description{
Produces a list of names of measures that can be used in CV for different
families. Note, however, that the package does not check if the measure
the user specifies is appropriate for the family.
}
