% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValidTypeMeasure.R
\name{checkValidTypeMeasure}
\alias{checkValidTypeMeasure}
\title{Check if loss function is valid for a given family}
\usage{
checkValidTypeMeasure(type.measure, family)
}
\description{
Also throws error if family is invalid
}
