% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCindex.R
\name{getCindex}
\alias{getCindex}
\title{Compute C index for a Cox model}
\usage{
getCindex(pred, y, weights = rep(1, nrow(y)))
}
\arguments{
\item{pred}{A vector of predictions.}

\item{y}{Survival response variable, must be a \code{Surv} or
\code{stratifySurv} object.}

\item{weights}{Observation weights (default is all equal to 1).}
}
\description{
Computes Harrel's C (concordance) index for predictions, taking
censoring into account.
}
