% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTypeMeasureName.R
\name{getTypeMeasureName}
\alias{getTypeMeasureName}
\title{Get full name of loss function}
\usage{
getTypeMeasureName(type.measure, family)
}
\arguments{
\item{type.measure}{Loss function to use for cross-validation.}

\item{family}{Model family.}
}
\value{
A named vector of length 1. The vector's value is the full name
of the loss function, while the name of that element is the short name
of the loss function.
}
\description{
Get the full name of the loss function from `type.measure` and `family`.
}
