% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cvobj.R
\name{plot.cvobj}
\alias{plot.cvobj}
\title{Plot the cross-validation curve from a class `cvobj` object}
\usage{
\method{plot}{cvobj}(x, sign.lambda = 1, log.lambda = TRUE, ...)
}
\arguments{
\item{x}{A `"cvobj"` object.}

\item{sign.lambda}{Either plot against `log(lambda)` (default) or its
negative if `sign.lambda = -1`.}

\item{log.lambda}{If `TRUE` (default), x-axis is `log(lambda)` instead of
`lambda` (`log.lambda = FALSE`).}

\item{...}{Other graphical parameters to plot.}
}
\value{
A plot is produced, and nothing is returned.
}
\description{
Plots the cross-validation curve, and upper and lower standard deviation
curves, as a function of the `lambda` values used.
}
