% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cvobj.R
\name{print.cvobj}
\alias{print.cvobj}
\title{Print a class `cvobj` object}
\usage{
\method{print}{cvobj}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A `"cvobj"` object.}

\item{digits}{Significant digits in printout.}

\item{...}{Other print arguments.}
}
\value{
A summary is printed, and nothing is returned.
}
\description{
Print a summary of results of cross-validation for a class `cvobj` object.
}
