% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeError.R
\name{computeError}
\alias{computeError}
\title{Compute CV statistics from a prediction matrix}
\usage{
computeError(
  predmat,
  y,
  lambda,
  foldid,
  type.measure,
  family,
  weights = rep(1, nrow(predmat)),
  grouped = TRUE
)
}
\arguments{
\item{predmat}{nobs by nlambda matrix of predictions. Note that these
should be on the same scale as `y` (unlike in the glmnet package where it
is the linear predictor).}

\item{y}{Response variable.}

\item{lambda}{Lambda values associated with the errors in `predmat`.}

\item{foldid}{Vector of values identifying which fold each observation is
in.}

\item{type.measure}{Loss function to use for cross-validation. See
`availableTypeMeasures()` for possible values for `type.measure`. Note that
the package does not check if the user-specified measure is appropriate
for the family.}

\item{family}{Model family; used to determine the correct loss function.}

\item{weights}{Observation weights.}
}
\description{
Compute CV statistics from a matrix of predictions.
}
