% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTypeMeasureName.R
\name{getTypeMeasureName}
\alias{getTypeMeasureName}
\title{Get full name of loss function}
\usage{
getTypeMeasureName(type.measure, family)
}
\description{
Get the full name of the loss function from `type.measure` and `family`.
}
